% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_scan_targeted.R
\name{annotate_scan_targeted}
\alias{annotate_scan_targeted}
\title{Annotating a deconvoluted oligonucleotide spectra}
\usage{
annotate_scan_targeted(
  scan_processed_aggregated,
  formula_flp = "C192H239O117N73P18S4F8",
  transformation_list = NULL,
  mdb = NULL,
  ntheo = 12
)
}
\description{
The function searches a complex deconvoluted oligonucleotide spectra against a user provided oligonucleotide impurity/metabolite database, annotating and scoring isotopic pattern matches. Alternatively, user can provide the FLP (full length product) formula and a list of bio or chemical transformations.
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
