% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_scan.R
\name{centroid_scan}
\alias{centroid_scan}
\title{Centroiding mass spectra}
\usage{
centroid_scan(scan0)
}
\description{
The function creates centroid-mode mass spectrum if profile mode mass spectrum is detected
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
