% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution1.R
\name{deconvolution1}
\alias{deconvolution1}
\title{Deconvoluting a mixed envelop to two pure components}
\usage{
deconvolution1(
  scan_df,
  theor_ID_cmpd1,
  theor_ID_cmpd2,
  n_theor_peaks,
  expected_charge_range,
  matching_mass_accuracy,
  noise_threshold,
  deduplicate_fun
)
}
\arguments{
\item{scan_df}{Data frame with two columns, mz and intensity, representing all peaks in a single mass scan at one selected retention time}

\item{n_theor_peaks}{Numeric. How many theoretical peaks should be returned by BRAIN}

\item{expected_charge_range}{Numeric, which charge state should be analysed? e.g. 5:11}

\item{matching_mass_accuracy}{Numeric, relative (in Da) mass tolerance for assigning observed peaks to theoretical peaks}

\item{noise_threshold}{Numeric, keep for the analysis only those peaks with intensities greater or equal than this threshold}

\item{deduplicate_fun}{Character, "max" by default. How to deduplicate multiple observed peaks assigned to one theoretical peak - "max" or "sum"}

\item{ac_cmpd1}{List used by the BRAIN algorithm to compute theoretical isotope distributions. The lists have to follow this format: list(C = 1, H = 2, F = 3, N = 4, O = 5, P = 6, S = 7)), where numbers 1-7 should be replaced with actual values}

\item{ac_cmpd2}{Same as previous input, another expected theoritical envelop list}
}
\value{
\itemize{
   \item{by_charge:}{ list, with elements like z5, z6, ..., each storing estimate (estimated proportion of cmpd2, 0-1 scale),
                   se (standard error of the proportion estimate), p_value (null hypothesis: estimate=0), 
                   significance (1 = significant, 0 = insignificant), mpcse (pearson chi-square goodness-of-fit, similar to the pointless4dna paper)}
   \item{by_charge:}{ not yet available}
 }
}
\description{
The function decomposes a mixed envelop into two theoritical isotope envelopes. It quantifies the two components by decomposing.
}
\author{
Piotr Prostko, \email{piotr.prostko@uhasselt.be}
}
