% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_coverage.R
\name{display_coverage}
\alias{display_coverage}
\title{Display sequence coverage}
\usage{
display_coverage(
  scan.deconvoluted.annotated,
  seq,
  int.frag = F,
  return.plot = T
)
}
\description{
The function returns a ggplot of labelled fragments and internal fragments on an oligonucleotide sequence
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
