% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_esi_frag_basic.R
\name{predict_esi_frag_basic}
\alias{predict_esi_frag_basic}
\title{Prediction ESI fragments first function}
\usage{
predict_esi_frag_basic(test_seq = "OH-Ad-Ad-Ad-Ad-Ad-Ad-OH")
}
\description{
The function creates fragment ions for a sequence given
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
