% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scan.R
\name{process_scan}
\alias{process_scan}
\title{Converting multi-charged oligonucleotide mass spectra to true molecular weight scale}
\usage{
process_scan(
  test.scan,
  polarity = c("Positive", "Negative"),
  MSMS = F,
  baseline = 100,
  min_charge = 3,
  max_charge = 12,
  min_mz = 500,
  max_mz = 1500,
  min_mw = 4000,
  max_mw = 12000,
  mz_error = 0.02,
  mw_gap = 5
)
}
\description{
The function determines charges state based on peak spacing. It also creates a deconvoluted molecular weight spectra by combining multiple charge state of the same isotopic species.
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
