% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scan_maldi.R
\name{process_scan_maldi}
\alias{process_scan_maldi}
\title{Converting maldi oligonucleotide mass spectra}
\usage{
process_scan_maldi(
  test.scan,
  polarity = c("Positive", "Negative"),
  baseline = 1000,
  min_mz = 100,
  max_mz = 2500,
  mw_gap = 1.1,
  mw_window = 10
)
}
\description{
The function processes and envelops maldi scan
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
