% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_scan_annotated.R
\name{reconstruct_scan_annotated}
\alias{reconstruct_scan_annotated}
\title{Reconstruct a theoretical mass spectra based on oligonucleotide features detected}
\usage{
reconstruct_scan_annotated(
  scan0,
  scans.deconvoluted.annotated,
  polarity = "Negative",
  baseline = 100,
  mode = c("targeted", "untargeted"),
  mz_error = 0.01,
  bblock = "C21 H26.4 O13.2 N7.3 P2 S0.4 F0.9",
  ntheo = 12
)
}
\description{
The function provides a way to check the detected features against an original spectrum.
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
