% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_scan_mix.R
\name{annotate_scan_mix}
\alias{annotate_scan_mix}
\title{Targeted followed by non-targeted screening from a deconvoluted oligonucleotide spectra}
\usage{
annotate_scan_mix(
  scan_processed_aggregated,
  MSMS = F,
  ntheo = 10,
  formula_flp = "C192H239O117N73P18S4F8",
  cpd_flp = "Demo A",
  transformation_list = NULL,
  mdb = NULL,
  bblock = "DNA",
  min_overlap = 0.6,
  max_msigma = 5,
  max_mmw_ppm = 10,
  baseline = 1000
)
}
\description{
The function first searches a complex deconvoluted oligonucleotide spectra against a user provided oligonucleotide impurity/metabolite database, annotating and scoring isotopic pattern matches. It then searches DNA/RNA-like isotope patterns from the rest of deconvoluted oligonucleotide spectra. It provides the monoisotopic molecular weight, average, intensity and envelope likeness of all features detected.
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
