% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_scan_untargeted.R
\name{annotate_scan_untargeted}
\alias{annotate_scan_untargeted}
\title{Non-targeted screening from a deconvoluted oligonucleotide spectra}
\usage{
annotate_scan_untargeted(
  scan_processed_aggregated,
  bblock,
  ntheo = 12,
  min_overlap = 0.6,
  max_msigma = 10,
  baseline = 1000
)
}
\description{
The function searches DNA/RNA-like isotope patterns from a deconvoluted oligonucleotide spectra. It provides the monoisotopic molecular weight, average, intensity and envelope likeness of all features detected.
}
\author{
Youzhong Liu, \email{YLiu186@ITS.JNJ.com}
}
